+++ internal/provider/resource_authenticator.go
@@ -230,16 +230,17 @@ func (r *ConjurAuthenticatorResource) Configure(ctx context.Context, req resourc
 	}
 	client, ok := req.ProviderData.(api.ClientV2)
 	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected api.ClientV2, got: %T", req.ProviderData),
-		)
+		AddUnexpectedConfigureTypeError(&resp.Diagnostics, "api.ClientV2", req.ProviderData)
 		return
 	}
 	r.client = client
 }
 
 func (r *ConjurAuthenticatorResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurAuthenticatorResourceModel
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -271,6 +272,10 @@ func (r *ConjurAuthenticatorResource) Create(ctx context.Context, req resource.C
 }
 
 func (r *ConjurAuthenticatorResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurAuthenticatorResourceModel
 
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
@@ -296,6 +301,10 @@ func (r *ConjurAuthenticatorResource) Read(ctx context.Context, req resource.Rea
 }
 
 func (r *ConjurAuthenticatorResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurAuthenticatorResourceModel
 
 	// Read Terraform plan data into the model
@@ -324,6 +333,10 @@ func (r *ConjurAuthenticatorResource) Update(ctx context.Context, req resource.U
 }
 
 func (r *ConjurAuthenticatorResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurAuthenticatorResourceModel
 
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)