+++ internal/provider/resource_group.go
@@ -118,16 +118,17 @@ func (r *ConjurGroupResource) Configure(ctx context.Context, req resource.Config
 	}
 	client, ok := req.ProviderData.(api.ClientV2)
 	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected api.ClientV2, got: %T", req.ProviderData),
-		)
+		AddUnexpectedConfigureTypeError(&resp.Diagnostics, "api.ClientV2", req.ProviderData)
 		return
 	}
 	r.client = client
 }
 
 func (r *ConjurGroupResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurGroupResourceModel
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -154,6 +155,10 @@ func (r *ConjurGroupResource) Create(ctx context.Context, req resource.CreateReq
 }
 
 func (r *ConjurGroupResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurGroupResourceModel
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
 	if resp.Diagnostics.HasError() {
@@ -165,6 +170,10 @@ func (r *ConjurGroupResource) Update(ctx context.Context, req resource.UpdateReq
 }
 
 func (r *ConjurGroupResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurGroupResourceModel
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
 
@@ -196,6 +205,10 @@ func (r *ConjurGroupResource) Read(ctx context.Context, req resource.ReadRequest
 }
 
 func (r *ConjurGroupResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurGroupResourceModel
 
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)