+++ internal/provider/resource_permission.go
@@ -149,16 +149,17 @@ func (r *ConjurPermissionResource) Configure(ctx context.Context, req resource.C
 	}
 	client, ok := req.ProviderData.(api.ClientV2)
 	if !ok {
-		resp.Diagnostics.AddError(
-			"Unexpected Resource Configure Type",
-			fmt.Sprintf("Expected api.ClientV2, got: %T", req.ProviderData),
-		)
+		AddUnexpectedConfigureTypeError(&resp.Diagnostics, "api.ClientV2", req.ProviderData)
 		return
 	}
 	r.client = client
 }
 
 func (r *ConjurPermissionResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurPermissionResourceModel
 
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
@@ -184,6 +185,10 @@ func (r *ConjurPermissionResource) Create(ctx context.Context, req resource.Crea
 }
 
 func (r *ConjurPermissionResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurPermissionResourceModel
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)
 
@@ -216,6 +221,10 @@ func (r *ConjurPermissionResource) Read(ctx context.Context, req resource.ReadRe
 }
 
 func (r *ConjurPermissionResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurPermissionResourceModel
 	resp.Diagnostics.Append(req.Plan.Get(ctx, &data)...)
 
@@ -236,6 +245,10 @@ func (r *ConjurPermissionResource) Update(ctx context.Context, req resource.Upda
 }
 
 func (r *ConjurPermissionResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	if r.client == nil {
+		AddProviderClientNotConfiguredWarning(&resp.Diagnostics)
+		return
+	}
 	var data ConjurPermissionResourceModel
 
 	resp.Diagnostics.Append(req.State.Get(ctx, &data)...)